+++ src/main/java/dev/lambdaurora/lambdynlights/DynamicLightsConfig.java
@@ -30,7 +30,7 @@ import java.util.List;
  * Represents the mod configuration.
  *
  * @author LambdAurora
- * @version 4.0.0
+ * @version 4.1.0
  * @since 1.0.0
  */
 public class DynamicLightsConfig {
@@ -53,6 +53,7 @@ public class DynamicLightsConfig {
 	private final BooleanSettingEntry selfLightSource;
 	private final BooleanSettingEntry waterSensitiveCheck;
 	private final BooleanSettingEntry beamLighting;
+	private final BooleanSettingEntry fireflyLighting;
 	private final BooleanSettingEntry guardianLaser;
 	private final BooleanSettingEntry debugActiveDynamicLightingCells;
 	private final BooleanSettingEntry debugDisplayDynamicLightingChunkRebuild;
@@ -94,6 +95,10 @@ public class DynamicLightsConfig {
 				"light_sources.beam", true, this.config,
 				Text.translatable("lambdynlights.option.light_sources.beam.tooltip")
 		);
+		this.fireflyLighting = new BooleanSettingEntry(
+				"light_sources.firefly", true, this.config,
+				Text.translatable("lambdynlights.option.light_sources.firefly.tooltip")
+		);
 		this.guardianLaser = new BooleanSettingEntry(
 				"light_sources.guardian_laser", true, this.config,
 				Text.translatable("lambdynlights.option.light_sources.guardian_laser.tooltip")
@@ -116,6 +121,7 @@ public class DynamicLightsConfig {
 				this.selfLightSource,
 				this.waterSensitiveCheck,
 				this.beamLighting,
+				this.fireflyLighting,
 				this.guardianLaser,
 				this.debugActiveDynamicLightingCells,
 				this.debugDisplayDynamicLightingChunkRebuild,
@@ -258,6 +264,13 @@ public class DynamicLightsConfig {
 	 */
 	public BooleanSettingEntry getBeamLighting() {
 		return this.beamLighting;
+	}
+
+	/**
+	 * {@return the firefly light source setting holder}
+	 */
+	public BooleanSettingEntry getFireflyLighting() {
+		return this.fireflyLighting;
 	}
 
 	/**